% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pct_groups.R
\name{calc_pct_groups}
\alias{calc_pct_groups}
\title{Funcion auxiliar para generacion de porcentaje por factores}
\usage{
calc_pct_groups(df, col, q, negocio)
}
\arguments{
\item{negocio}{}
}
\description{
Funcion auxiliar para generacion de porcentaje por factores
}
