% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBaseIPN.R
\name{calcBaseIPN}
\alias{calcBaseIPN}
\title{Funcion aux: Cuenta renglones distintos a NA}
\usage{
calcBaseIPN(df, c)
}
\arguments{
\item{c}{Columna de tipo String}
}
\value{
Número de filas que no sean NA
}
\description{
Esta función cuenta el número de entradas distintas a NA que hay en la columna `c` en la tabla `df`
}
