% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIPN.R
\name{calcIPN}
\alias{calcIPN}
\title{Calculo del IPN}
\usage{
calcIPN(df, column)
}
\arguments{
\item{df}{Base de datos}

\item{column}{Columna de calificaciones en formato string.}
}
\description{
Esta funci칩n calcula el IPN con base en la columna cuyo nombre est치 en la
variable `lab` (la cual puede contener el nombre de la columna en
tipo `string` o `symbol`), que tiene clasificados a los usuarios
como detractor, pasivo o promotor. Esta funci칩n se utiliza principalmente para
imprimir el IPN dentro de la gr치fica de anillo, pero puede utilizarse para
calcular IPNs en general, mientras se tenga la tabla una columna que
clasifique a los usuarios con niveles "Detractor", "Pasivo" y "Promotor".
}
