% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{cleaning}
\alias{cleaning}
\title{Función auxiliar de limpieza para grafica grid}
\usage{
cleaning(data, cols, filt = NULL)
}
\arguments{
\item{filt}{}
}
\value{
DF con las columnas limpias
}
\description{
Esta funcion filtra la base y convierte las preguntas de escala 0 a 10 a nunéricas
}
