% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundir_hist.R
\name{fundir_hist}
\alias{fundir_hist}
\title{Función auxiliar que fundo una tabla de ion historico}
\usage{
fundir_hist(dfh, negocio)
}
\arguments{
\item{negocio}{}
}
\value{
df
}
\description{
Función auxiliar que fundo una tabla de ion historico
}
