% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevelsTable.R
\name{getLevelsTable}
\alias{getLevelsTable}
\title{Funcion aux para tabla menciones}
\usage{
getLevelsTable(df, variable)
}
\arguments{
\item{variable}{}
}
\description{
Funcion aux para tabla menciones
}
