% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevelsWTotal.R
\name{getLevelsWTotal}
\alias{getLevelsWTotal}
\title{Funcion aux para tabla menciones}
\usage{
getLevelsWTotal(df, variable)
}
\arguments{
\item{variable}{}
}
\description{
Funcion aux para tabla menciones
}
