% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipn_hist_tipo_plot.R
\name{ipn_hist_tipo_plot}
\alias{ipn_hist_tipo_plot}
\title{Genera grafica de ipn historico por tipo cliente}
\usage{
ipn_hist_tipo_plot(data)
}
\arguments{
\item{data}{}
}
\description{
Genera grafica de ipn historico por tipo cliente
}
