% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_historico_plot.R
\name{make_historico_plot}
\alias{make_historico_plot}
\title{Generador tabla IPN historico}
\usage{
make_historico_plot(ipn_hist, Q_s)
}
\arguments{
\item{ipn_hist}{Eje y}

\item{Q_s}{Labels en el eje x}
}
\value{
plotly plot
}
\description{
Entran unicamente labels
}
