% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBar.R
\name{makeBar}
\alias{makeBar}
\title{Grafica de barras horizontales para IPN}
\usage{
makeBar(df, label_var, level_var, filtros = NULL)
}
\arguments{
\item{filtros}{}
}
\value{
Desripción pendiente
}
\description{
Desripción pendiente
}
