% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBar_hist.R
\name{makeBar_hist}
\alias{makeBar_hist}
\title{Grafica de barras horizontales para IPN con diferencia del Q pasado}
\usage{
makeBar_hist(df, label_var, level_var, df_ant, lever_var_ant,
  filtros = NULL)
}
\arguments{
\item{filtros}{}
}
\value{
Plotly plot
}
\description{
Grafica de barras horizontales para IPN con diferencia del Q pasado
}
