% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMenciones.R
\name{makeMenciones}
\alias{makeMenciones}
\title{Generador de tabla de menciones}
\usage{
makeMenciones(df, menciones_col, tipo = "Todas", cruce = "Ninguno",
  filtro = "Ninguno", filtro_col = "")
}
\arguments{
\item{filtro_col}{}
}
\description{
Generador de tabla de menciones
}
