% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRing.R
\name{makeRing}
\alias{makeRing}
\title{Grafica de disco para IPN}
\usage{
makeRing(df, column, title = "", textsize = 16, colors = c("cc3232",
  "e7b416", "298428"))
}
\arguments{
\item{df}{Base de datos}

\item{column}{Columna donde se encuentran las calificaciones para el calculo del IPN}

\item{title}{Titulo de la gráfica, argumento se le pasa a plotly}

\item{textsize}{Tamaño del texto, argumento se le pasa a plotly}

\item{colors}{Colores de promoteres, pasivos y detractores, argumento se le pasa a plotly}
}
\value{
Objeto tipo plot_ly
}
\description{
Esta función genera la gráfica de anillo de los porcentajes correspondientes
a los niveles de la columna `column` de la tabla `df`, ignorando NAs.
La gráfica se genera utilizando plotly.
Esta función ignora NAs, y presupone que se van a graficar tres niveles:
"Detractor", "Pasivo" y "Promotor", en ese orden, coloréandolos
de rojo, amarillo y verde, respectivamente.
Es posible graficar la frecuencia de un mayor número de niveles, pero
plotly escogerá los colores de manera automática, a menos de que se
especifiquen en los parámetros de la función.
}
