% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menciones_comp.R
\name{menciones_comp}
\alias{menciones_comp}
\title{Tabla de menciones con comparativo al Q anterior}
\usage{
menciones_comp(df, df_old, menciones_col, menciones_col_old, q,
  tipo = "Todas", cruce = "Ninguno", filtro = "Ninguno",
  filtro_col = "", filtro_col_old = filtro_col)
}
\arguments{
\item{df}{Base de datos actual}

\item{df_old}{Base de datos del Q pasado}

\item{menciones_col}{Vector con los nombres de las columnas donde están las menciones de `df`}

\item{menciones_col_old}{Vector con los nombres de las columnas donde están las menciones de `df_old`}

\item{q}{Número de levantamiento (i.e. 3Q)}

\item{tipo}{Tipo de menciones que se quieren ("Positivas", "Negativas" o "Todas")}

\item{cruce}{Columna con la cual se van a cruzar las menciones}

\item{filtro}{Tipos de clientes: "Promotor", "Pasivo", "Detractor" o "Pasivo + Detractor"}

\item{filtro_col}{Columna en `df` en la cual está el filtro buscado}

\item{filtro_col_old}{Columna en `df_old` en la cual está el filtro buscado}
}
\value{
Tabla con columnas: mención, casos, porcentaje actual, porcentaje pasado, diferencia y código html para imprimir
}
\description{
Esta función recibe los datos del actual Q y el Q anterior y con ello calcula
el la frecuencia de menciones y porcentajes de éstas. Luego realiza un left_join
con fuzzymatching (utilizando distancia de texto coseno, modificable) por la izquierda,
y con ello saca las diferencias porcentuales del Q anterior a este.
Esta función es una versión más elaborada de makeMenciones y la manda a llamar
para su funcionamiento.
}
