% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\title{Data prep para la grafica grid}
\usage{
prep(data, ipn, prob)
}
\arguments{
\item{prob}{}
}
\description{
this function and the two following it create the intermediate tables function 'makeTiles' needs
"ipn" is the column name for the NPS question, and 'prob' is the name of the column containing the question
"Con qué probabilidad seguiría siendo cliente de XYZ', o, 'con qué probabilidad volvería a obtener un crédito etc'
}
