% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepros.R
\name{prepros}
\alias{prepros}
\title{Función auxiliar para generación de la gráfica de barras}
\usage{
prepros(df, label_var, level_var, filtros = NULL)
}
\arguments{
\item{filtros}{}
}
\value{
Descripción pendiente
}
\description{
Descripción pendiente
}
