% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_q.R
\name{previous_q}
\alias{previous_q}
\title{Funcion previous_q:
Dado un entero n (1 <= n <= 4) se obtiene n-1 correspondiente al trimestre (Q) anterior (módulo 4)}
\usage{
previous_q(n)
}
\arguments{
\item{n}{Entero entre 1 y 4}
}
\value{
n-1 (mod 4)
}
\description{
Funcion previous_q:
Dado un entero n (1 <= n <= 4) se obtiene n-1 correspondiente al trimestre (Q) anterior (módulo 4)
}
